students/k33422/Malaev_Stepan/lr_2/pyproject.toml
[project]
name = "lr-2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.5",
    "asyncpg>=0.29.0",
    "psycopg2>=2.9.9",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
